Add actionlint Git hook (#12)
# Pull Request

## Description

This pull request introduces a new `actionlint` check to the project by
adding a corresponding task to the `Justfile` and integrating it into
the `lefthook` pre-commit hooks. This ensures that GitHub Actions
workflows are validated for potential issues during development.

### New `actionlint` integration:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR49-R56):
Added a new `actionlint-check` task to run `actionlint` checks. This
task is now part of the available commands for validating GitHub Actions
workflows.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR23-R24):
Integrated the `actionlint-check` task into the `pre-commit` hooks under
the name "Actionlint Checks," ensuring it runs automatically before
commits.