Update Project Documentation (#16)
# Pull Request

## Description

This pull request updates the repository to reflect its new name, "One
Time Scripts," by making changes to the `README.md` and
`docs/SECURITY.md` files. The updates ensure consistency in
documentation and links.

### Repository renaming updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R5):
Updated the repository name in the title and Table of Contents to "One
Time Scripts."
*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L9-R9):
Updated links in the "Security tab" and "Security Advisories" sections
to point to the renamed repository, "one-time-scripts."
[[1]](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L9-R9)
[[2]](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L23-R23)